# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NIX Platform API
  description: |
    API REST do NIX Platform — um monólito modular que centraliza
    integrações, automação e notificações. Todo endpoint autenticado espera
    um access token OIDC emitido pelo Keycloak como Bearer JWT; o backend o
    valida localmente contra um JWKS em cache (nunca com uma chamada por
    requisição ao Keycloak — ver §29 do documento de arquitetura).

    Toda resposta usa o mesmo envelope:
    `{"data": <payload|null>, "error": {"code","message"}|null, "meta": {...}|omitido}`.
    Erros nunca incluem stack trace ou outro detalhe interno.

    **Idempotência**: qualquer endpoint de escrita (POST/PUT/PATCH) aceita
    opcionalmente o header `Idempotency-Key` (ver
    `#/components/parameters/IdempotencyKey`). Reenviar a mesma requisição
    (mesmo método, caminho e corpo) com a mesma chave nunca executa o
    handler duas vezes — a segunda chamada em diante recebe de volta a
    resposta já produzida pela primeira (`Idempotent-Replay: true` no
    header da resposta), em vez de criar um segundo job/efeito colateral.
    Reusar a mesma chave com um corpo diferente é rejeitado com 409
    (`IDEMPOTENCY_KEY_REUSED`); uma segunda chamada enquanto a primeira
    ainda está em andamento recebe 409 (`IDEMPOTENCY_KEY_IN_PROGRESS`). Ver
    ADR [002](adr/002-enterprise-resilience-and-governance.md).

    **Resiliência e disponibilidade condicional**: chamadas que dependem
...
  version: "1.0.0"
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/scanning/scans:
    post:
      tags: [scanning]
      summary: Dispara um scan assíncrono contra um alvo
      operationId: createScan
      description: >
        Não executa o(s) scanner(s) em si. Cria um job, o persiste junto com
        um evento de outbox em uma única transação, e retorna imediatamente
        — o worker o processa (clona o alvo via git e roda cada scanner
        pedido), e o frontend é notificado via WebSocket quando ele
        concluir ou falhar. `job_id` retornado é o mesmo `scan_id` usado em
        `GET /api/v1/scanning/scans/{scanID}/findings`. Sujeito a rate
        limiting por chamador. Aceita `Idempotency-Key`.

        `scanners` aceita mais de um nome — todos rodam em PARALELO contra
        o mesmo alvo, sob o mesmo job/`scan_id` (Fase 7 — Orquestração
        concorrente). Um scanner que falha não afeta os demais: o job é
        marcado concluído se PELO MENOS UM tiver sucesso (achados só dos
        que tiveram sucesso; os que falharam ficam registrados no
        resultado do job e na auditoria) — só falha por completo, sujeito
        a retry, se TODOS falharem.
      security:
        - bearerAuth: [scanning:manage]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scanners, target]
              properties:
                scanners:
                  type: array
                  minItems: 1
                  description: Um ou mais nomes de CodeScanner registrado — todos rodam em paralelo.
                  items:
                    type: string
                    enum: ["trivy", "semgrep", "sonarqube", "zap"]
                  example: ["trivy", "semgrep"]
                target:
                  type: string
                  description: >
                    Para "trivy" (dependências, Dockerfiles), "semgrep"
                    (SAST) e "sonarqube" (qualidade de código/bugs) — os
                    três clonam o alvo via git antes de escanear: uma URL
                    git https://, com um branch/tag opcional após "#"
                    (ex.: "https://github.com/org/repo.git#main"). Só
                    https:// é aceito; o host é resolvido e rejeitado se
                    apontar para um IP privado/loopback/link-local (defesa
                    de SSRF, ver docs/roadmap-secops-orchestrator.md). O
                    scanner "sonarqube" exige SCANNING_SONARQUBE_URL
                    configurada no worker (503 DEPENDENCY_UNAVAILABLE caso
                    contrário) e leva mais tempo — além do scan em si, o
                    worker espera a Compute Engine do servidor terminar de
                    processar o relatório enviado.

                    Para "zap" (DAST) — completamente diferente: o alvo é
                    uma URL http(s) de um serviço RODANDO DE VERDADE (não
                    um repositório git), que o ZAP ativamente ataca
                    (spider + scan ativo — injeção, XSS, etc.). Exige que
                    o host do alvo esteja na allowlist configurada em
                    SCANNING_ZAP_ALLOWED_HOSTS — vazia por padrão, recusa
                    TODO alvo até um host de staging/homologação ser
                    explicitamente autorizado (400 VALIDATION_ERROR caso
                    contrário). Nunca aponte para produção.
                  example: "https://github.com/org/repo.git#main"
      responses:
        "202":
          $ref: "#/components/responses/TestJobAccepted"
        "400":
          description: Corpo inválido (target vazio, JSON malformado).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Um ou mais nomes em `scanners` não estão registrados — nenhum job é criado nesse caso, nem para os nomes válidos.
          content:
            application/json:
              example:
                data: null
                error: { code: "NOT_FOUND", message: "scanner(s) not registered: does-not-exist" }
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/scanning/findings:
    get:
      tags: [scanning]
      summary: Lista os achados mais graves/recentes entre TODAS as execuções de scan
      operationId: listRecentFindings
      description: >
        Diferente de `GET /api/v1/scanning/scans/{scanID}/findings` (escopado a um scan_id já
        conhecido), este endpoint não exige saber um scan_id de antemão — o feed usado pela UI
        (Fase 9 do roadmap de segurança) pra listar achados recentes por severidade.
      security:
        - bearerAuth: [scanning:read]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: limit
          in: query
          required: false
          description: >
            Alias de page_size, aceito por compatibilidade com chamador anterior à Fase 14
            (Maturidade de AppSec) — só usado quando page_size não vem explícito. Prefira page_size
            em código novo.
          schema:
            type: integer
      responses:
        "200":
          description: >
            Uma PÁGINA de achados, mais graves/recentes primeiro (lista vazia se nenhum scan rodou
            ainda). Antes da Fase 14, este endpoint tinha um `limit` sem `OFFSET` (teto fixo de 200)
            — qualquer achado além disso nunca aparecia; agora é paginação de verdade, mesmo
            contrato que GET /api/v1/users já usa.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Finding"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/scanning/projects/{projectID}/findings-history:
    get:
      tags: [scanning]
      summary: Histórico de achados deduplicado de um projeto (Fase 12)
      operationId: listProjectFindingsHistory
      description: >
        Um achado por fingerprint, não por linha de scan_findings — "apareceu pela primeira vez no
        scan de 12/08, ainda presente no scan de 20/08" em vez de listar a mesma vulnerabilidade uma
        vez por re-scan. Nunca documentado antes desta revisão (Fase 14), apesar de existir desde a
        Fase 12 — lacuna da documentação, não do endpoint em si.
      security:
        - bearerAuth: [scanning:read]
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Histórico deduplicado, achado aberto e não triado primeiro (lista vazia se o projeto nunca rodou um scan).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProjectFindingHistory"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/scanning/projects/{projectID}/findings/{fingerprint}/triage:
    put:
      tags: [scanning]
      summary: Tria um achado deduplicado (Fase 14 — Maturidade de AppSec)
      operationId: triageFinding
      description: >
        Marca o fingerprint como false_positive/wont_fix/risk_accepted dentro do projeto — chamar de
        novo no MESMO fingerprint substitui a decisão anterior, não acumula histórico. reason é
        obrigatório: uma supressão sem justificativa registrada é o tipo de decisão que uma auditoria
        de segurança depois cobra explicação. Gravado em audit_logs (finding.triaged).
      security:
        - bearerAuth: [scanning:manage]
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, reason]
              properties:
                status:
                  type: string
                  enum: [false_positive, wont_fix, risk_accepted]
                reason:
                  type: string
                  description: Obrigatório — motivo da triagem, no máximo 2000 caracteres.
                expires_at:
                  type: string
                  format: date-time
                  description: >
                    Opcional (Fase 14, continuação — expiração de triagem). Precisa estar no futuro.
                    Quando o prazo passa, o achado volta a contar como aberto em
                    SecurityPosture/ListProjectFindingsHistory (mas a decisão em si nunca é apagada
                    automaticamente — ver GET .../findings-history's triage_expired).
      responses:
        "200":
          description: Triagem gravada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: projectID/fingerprint inválido, status desconhecido, reason vazio/longo demais, ou expires_at no passado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Projeto não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      tags: [scanning]
      summary: Reabre um achado já triado (Fase 14 — Maturidade de AppSec)
      operationId: untriageFinding
      description: >
        Apaga a decisão de triagem — idempotente, chamar num fingerprint nunca triado (ou já
        reaberto) não é erro. Gravado em audit_logs (finding.untriaged).
      security:
        - bearerAuth: [scanning:manage]
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Achado reaberto (ou já estava aberto).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: projectID/fingerprint inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Um access token OIDC emitido pelo Keycloak. Validado localmente
        (assinatura, emissor, audiência, expiração, algoritmo) contra um
        JWKS em cache.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Chave única por operação lógica (recomenda-se um UUID gerado no
        cliente), escopada ao usuário autenticado. Reenviar a mesma
        requisição com a mesma chave reproduz a resposta original em vez
        de reexecutar a operação — ver a nota de Idempotência na descrição
        geral da API.
      schema:
        type: string
      example: "3f2b9e0a-6d34-4e77-9c3e-2b6f8e6b6a3d"
  responses:
    TestJobAccepted:
      description: >
        O teste foi enfileirado. Se a requisição carregou um
        `Idempotency-Key` já usado com sucesso antes (mesmo corpo), esta é
        uma resposta reproduzida (replay) da chamada original — o header
        `Idempotent-Replay: true` distingue esse caso, e nenhum job novo
        foi criado.
      headers:
        Idempotent-Replay:
          description: Presente e "true" somente quando esta resposta é um replay de uma chamada anterior com a mesma Idempotency-Key.
          schema:
            type: string
            enum: ["true"]
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      job_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [queued]
          example:
            data: { job_id: "3c1f1e2a-...", status: "queued" }
            error: null
    IdempotencyConflict:
      description: >
        Conflito relacionado à Idempotency-Key: `IDEMPOTENCY_KEY_IN_PROGRESS`
        se uma requisição com a mesma chave ainda está sendo processada, ou
        `IDEMPOTENCY_KEY_REUSED` se a chave já foi usada com um corpo de
        requisição diferente.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          examples:
            in_progress:
              summary: Requisição original ainda em andamento
              value:
                data: null
                error: { code: "IDEMPOTENCY_KEY_IN_PROGRESS", message: "a request with this Idempotency-Key is already being processed" }
            reused:
              summary: Mesma chave usada com um corpo diferente
              value:
                data: null
                error: { code: "IDEMPOTENCY_KEY_REUSED", message: "Idempotency-Key was already used with a different request" }
...
    Unauthorized:
      description: Access token ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          example:
            data: null
            error: { code: "UNAUTHORIZED", message: "invalid or expired access token" }
    Forbidden:
      description: Autenticado, mas sem o role/permissão necessário.
      content:
        application/json:
          example:
            data: null
            error: { code: "FORBIDDEN", message: "insufficient permission" }
...
    TooManyRequests:
      description: Limite de requisições excedido.
      content:
        application/json:
          example:
            data: null
            error: { code: "RATE_LIMITED", message: "too many requests, please try again later" }
  schemas:
    Envelope:
      type: object
      description: >
        O formato base de toda resposta. Operações bem-sucedidas restringem
        `data` a um tipo concreto via `allOf` no próprio schema da resposta;
        esta forma base documenta o contrato genérico {data, error, meta}.
      properties:
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ErrorBody"
        meta:
          type: object
          nullable: true
      required: [data, error]
...